SSE-2323: Move the session data declaration to its own file to allow reuse
Import the declaration as a module in the tests that need it instead of repeating the declaration.
