+++ express/tests/middleware/sign-in-status.test.ts
@@ -1,21 +1,7 @@
-import {AuthenticationResultType} from "@aws-sdk/client-cognito-identity-provider";
 import {NextFunction, Request, Response} from "express";
 import {Session, SessionData} from "express-session";
 import setSignInStatus from "../../src/middleware/sign-in-status";
-import {MfaResponse} from "../../src/services/self-service-services-service";
-
-declare module "express-session" {
-    interface SessionData {
-        emailAddress: string;
-        mobileNumber: string;
-        enteredMobileNumber: string;
-        cognitoSession: string;
-        authenticationResult: AuthenticationResultType;
-        mfaResponse: MfaResponse;
-        updatedField: string;
-        isSignedIn: boolean;
-    }
-}
+import "../../src/types/session";
 
 describe("Sign-in status is correctly set in res.locals so it can be accessed on any page", () => {
     let mockRequest: Partial<Request>;
