+++ express/tests/middleware/validators/emailIsPresentInSession.test.ts
@@ -1,21 +1,7 @@
-import {AuthenticationResultType} from "@aws-sdk/client-cognito-identity-provider";
 import {NextFunction, Request, Response} from "express";
 import {Session, SessionData} from "express-session";
 import emailIsPresentInSession from "../../../src/middleware/validators/emailIsPresentInSession/emailIsPresentInSession";
-import {MfaResponse} from "../../../src/services/self-service-services-service";
-
-declare module "express-session" {
-    interface SessionData {
-        emailAddress: string;
-        mobileNumber: string;
-        enteredMobileNumber: string;
-        cognitoSession: string;
-        authenticationResult: AuthenticationResultType;
-        mfaResponse: MfaResponse;
-        updatedField: string;
-        isSignedIn: boolean;
-    }
-}
+import "../../../src/types/session";
 
 describe("It checks whether an email is present in the session and behaves accordingly", () => {
     let mockRequest: Partial<Request>;
