+++ express/tests/middleware/validators/mobileOtpValidator.test.ts
@@ -1,21 +1,7 @@
-import {AuthenticationResultType} from "@aws-sdk/client-cognito-identity-provider";
 import {NextFunction, Request, Response} from "express";
 import {Session, SessionData} from "express-session";
 import {mobileSecurityCodeValidator} from "../../../src/middleware/validators/mobileOtpValidator";
-import {MfaResponse} from "../../../src/services/self-service-services-service";
-
-declare module "express-session" {
-    interface SessionData {
-        emailAddress: string;
-        mobileNumber: string;
-        enteredMobileNumber: string;
-        cognitoSession: string;
-        authenticationResult: AuthenticationResultType;
-        mfaResponse: MfaResponse;
-        updatedField: string;
-        isSignedIn: boolean;
-    }
-}
+import "../../../src/types/session";
 
 describe("It checks whether a mobile security code is valid and behaves accordingly", () => {
     let mockRequest: Partial<Request>;
